extract clashing property into separate enum
