+++ src/main/java/uk/gov/homeoffice/digital/sas/timecard/enums/ClashingProperty.java
@@ -1,0 +1,17 @@
+package uk.gov.homeoffice.digital.sas.timecard.enums;
+
+public enum ClashingProperty {
+  START_TIME("startTime"),
+  END_TIME("endTime"),
+  START_AND_END_TIME("startAndEndTime");
+  private final String stringValue;
+
+  ClashingProperty(final String s) {
+    stringValue = s;
+  }
+
+  @Override
+  public String toString() {
+    return stringValue;
+  }
+}
